servers-extra/weather/pyproject.toml
[project]
name = "mcp_server_weather"
version = "0.1.0"
description = "Weather MCP Server using MCP framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "typer>=0.9.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0"
]
dev = [
    "pytest>=7.0.0"
]

[project.scripts]
weather-server = "mcp_server_weather.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.package-data]
mcp_server_weather = ["**/*"]

[tool.setuptools]
packages = {find = {where = ["src"]}}

# 测试命令: uv run python -m pytest tests -v
